wait for replay file to exist before renaming
