+++ src/LiveSplit.OBSEvents/OBS/Protocol/Requests/Outputs/StartReplayBuffer.cs
@@ -1,7 +1,7 @@
-﻿using LiveSplit.OBSEvents.OBS.Protocol.Responses;
+﻿using LiveSplit.OBSEvents.OBS.Protocol.Responses.Outputs;
 using System;
 
-namespace LiveSplit.OBSEvents.OBS.Protocol.Requests
+namespace LiveSplit.OBSEvents.OBS.Protocol.Requests.Outputs
 {
     internal class StartReplayBuffer : Request<StartReplayBufferResponse>
     {
