+++ src/LiveSplit.OBSEvents/OBS/Protocol/Responses/Outputs/GetLastReplayBufferReplayResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace LiveSplit.OBSEvents.OBS.Protocol.Responses
+﻿namespace LiveSplit.OBSEvents.OBS.Protocol.Responses.Outputs
 {
     internal class GetLastReplayBufferReplayResponse(string requestId, RequestStatus requestStatus, string savedReplayPath) : RequestResponse(requestId, requestStatus)
     {
@@ -6,7 +6,7 @@
 
         public static GetLastReplayBufferReplayResponse Parse(dynamic json)
         {
-            dynamic data = ExtractResponseData(json);
+            dynamic data = ExtractAndValidateData(json);
             return Transform(data);
         }
